Rename targets.py -> target_types.py
